Adding api method for details
